product/serializers/category_serializer.py
from rest_framework import serializers

from product.models.category import Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = [
            'title',
            'sulg',
            'description',
            'active',
        ]